refactor: add new folders to organize components
