+++ src/components/feature/FilterInput.vue
@@ -1,0 +1,23 @@
+<template>
+  <div class="filter">
+    <input type="text"/>
+    <input type="text"/>
+    <input type="text"/>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'FilterInput',
+}
+</script>
+
+<style scoped>
+.card {
+  background-color: #FFFFFF;
+  border-radius: 10px;
+  box-shadow: 1px 1px 1px 2px #DDDDDD;
+  padding: 20px;
+  min-width: 580px;
+}
+</style>
